requirements.txt
snowflake-connector-python>=3.0.0
pyyaml>=5.4
dbt-core>=1.0.0,<2.0.0
dbt-snowflake>=1.0.0,<2.0.0